refactor: 将 editor 组件移出 moment
