+++ src/components/editor/text.tsx
